PDF to Audiobook Converter/requirements.txt
edge_tts==6.1.15
Pillow==11.0.0
PyPDF2==3.0.1
